requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
pmdarima>=2.0.0
prophet>=1.1.0
scipy>=1.9.0
python-dotenv>=0.19.0
answerrocket-sdk>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0